Move EditableSection to common
