+++ amundsen_application/static/js/components/common/EditableSection/styles.scss
